requirements.txt
absl-py==2.4.0
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
deprecated==1.3.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.15.1
idna==3.11
img2pdf==0.6.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
markdown==3.10.1
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.5.4
mouseinfo==0.1.3
namex==0.1.0
numpy==2.4.2
opencv-python==4.13.0.92
opt-einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pikepdf==10.3.0
pillow==12.1.0
protobuf==6.33.5
pyautogui==0.9.54
pydicom==3.0.1
pygetwindow==0.0.9
pygments==2.19.2
pymsgbox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
pyrect==0.2.0
pyscreeze==1.0.1
python-dateutil==2.9.0.post0
python-xlib==0.33
pytweening==1.2.0
reportlab==4.4.9
requests==2.32.5
rich==14.3.2
setuptools==81.0.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tkcap==0.0.4
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.1
pytest>=8.0,<9